redefine module at compile time, make utils coverage deterministc
