+++ apps/neoprice/mix.exs
@@ -10,8 +10,10 @@ defmodule Neoprice.Mixfile do
       deps_path: "../../deps",
       lockfile: "../../mix.lock",
       elixir: "~> 1.6",
+      elixirc_paths: elixirc_paths(Mix.env()),
       elixirc_options: [
-        warnings_as_errors: true
+        warnings_as_errors: true,
+        ignore_module_conflict: true
       ],
       start_permanent: Mix.env() == :prod,
       test_coverage: [
@@ -26,6 +28,10 @@ defmodule Neoprice.Mixfile do
       deps: deps()
     ]
   end
+
+  # Specifies which paths to compile per environment.
+  defp elixirc_paths(:test), do: ["lib", "test/support"]
+  defp elixirc_paths(_), do: ["lib"]
 
   # Run "mix help compile.app" to learn about applications.
   def application do
